Move backend's components to its own file.
